Move avatar event handlers into avatarEvents
