+++ behaviors/croquet/avatarEvents.js
@@ -1,3 +1,5 @@
+// import { GetPawn, q_yaw } from "@croquet/worldcore-kernel";
+
 class AvatarEventHandlerPawn {
     setup() {
         if (!this.isMyPlayerPawn) {return;}
@@ -28,6 +30,111 @@ class AvatarEventHandlerPawn {
 
         this.addLastResponder("keyUp", {ctrlKey: true}, this);
         this.addEventListener("keyUp", this.keyUp);
+    }
+
+    pointerTap(_e) {
+        if (this.editPawn) { // this gets set in pointerDown
+            this.editPawn.unselectEdit();
+            this.editPawn.showControls({avatar: this.actor.id,distance: this.targetDistance});
+            this.editPawn = null;
+            this.editPointerId = null;
+        }
+    }
+
+    pointerDown(e) {
+        if (e.ctrlKey || e.altKey) { // should be the first responder case
+            const render = this.service("ThreeRenderManager");
+            const rc = this.pointerRaycast(e.xy, render.threeLayerUnion('pointer'));
+            this.targetDistance = rc.distance;
+            let p3e = this.pointerEvent(rc, e);
+            p3e.lookNormal = this.actor.lookNormal;
+            let pawn = this.actor.worldcoreKernel.GetPawn(p3e.targetId);
+            pawn = pawn || null;
+
+            if (this.editPawn !== pawn) {
+                if (this.editPawn) {
+                    console.log('pointerDown clear old editPawn')
+                    this.editPawn.unselectEdit();
+                    this.editPawn = null;
+                    this.editPointerId = null;
+                }
+                console.log('pointerDown set new editPawn', pawn)
+                if (pawn) {
+                    this.editPawn = pawn;
+                    this.editPointerId = e.id;
+                    this.editPawn.selectEdit();
+                    this.buttonDown = e.button;
+                    if (!p3e.normal) {p3e.normal = this.actor.lookNormal}
+                    this.p3eDown = p3e;
+                }
+            } else {
+                console.log("pointerDown in editMode");
+            }
+        } else {
+            if (!this.focusPawn) {
+                // because this case is called as the last responder, facusPawn should be always empty
+                this.dragWorld = this.xy2yp(e.xy);
+                this.lookYaw = this.actor.worldcoreKernel.q_yaw(this._rotation);
+            }
+            let handlerModuleName = this.actor._cardData.avatarEventHandler;
+            this.call(`${handlerModuleName}$AvatarPawn`, "handlingEvent", "pointerDown", this, e);
+        }
+    }
+
+    pointerMove(e) {
+        if (this.editPawn) {
+            // a pawn is selected for draggging
+            if (e.id === this.editPointerId) {
+                if (this.buttonDown === 0) {
+                    this.editPawn.dragPlane(this.setRayCast(e.xy), this.p3eDown);
+                }else if (this.buttonDown == 2) {
+                    this.editPawn.rotatePlane(this.setRayCast(e.xy), this.p3eDown);
+                }
+            }
+        }else {
+            // we should add and remove responders dynamically so that we don't have to check things this way
+            if (!this.focusPawn && this.isPointerDown) {
+                let yp = this.xy2yp(e.xy);
+                let yaw = (this.lookYaw + (this.dragWorld[0] - yp[0]) * this.yawDirection);
+                let pitch = this.lookPitch + this.dragWorld[1] - yp[1];
+                pitch = pitch > 1 ? 1 : (pitch < -1 ? -1 : pitch);
+                this.dragWorld = yp;
+                this.lookTo(pitch, yaw);
+            }
+        }
+    }
+
+    pointerUp(_e) {
+        if (this.editPawn) {
+            this.editPawn.unselectEdit();
+            this.editPawn = null;
+            this.editPointerId = null;
+            this.p3eDown = null;
+            this.buttonDown = null;
+        }
+
+        // Below is a workaround to support an incomplete user program.
+        // If there are left over first responders (pointer capture) from a user object,
+        // delete them here.
+        if (this.firstResponders) {
+            for (let [_eventType, array] of this.firstResponders) {
+                for (let i = array.length - 1; i >= 0; i--) {
+                    let obj = array[i];
+                    if (obj.pawn !== this) {
+                        array.splice(i, 1);
+                    }
+                }
+            }
+        }
+    }
+
+    pointerWheel(e) {
+        let z = this.lookOffset[2];
+        z += Math.max(1,z) * e.deltaY / 1000.0;
+        z = Math.min(100, Math.max(z,0));
+        this.lookOffset = [this.lookOffset[0], z, z];
+        let pitch = (this.lookPitch * 11 + Math.max(-z / 2, -Math.PI / 4)) / 12;
+        this.lookTo(pitch, this.actor.worldcoreKernel.q_yaw(this._rotation), this.lookOffset); //,
     }
 
     startMotion(dx, dy) {
